pyproject.toml
[project]
name = "amazing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  	"Pillow>=10.0.0",
  	"Mlx",
	"pydantic"
]

[tool.setuptools.packages.find]
where = [".", "src"]

[tool.uv.sources]
mlx = { path = "mlx-2.2-py3-none-any.whl" }
